add(Repository): add an auth repository
